requirements.txt
pyspark==3.5.0
delta-spark==3.2.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
litellm>=1.0.0
pydantic>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
streamlit>=1.30.0
pytest>=8.0.0
faker>=25.0.0
python-dotenv>=1.0.0
requests>=2.31.0
duckdb>=0.10.0
google-generativeai>=0.4.0
